requirements.txt
python-binance
pandas
numpy
python-dotenv
matplotlib
scikit-learn
xgboost
